+++ packages/categories/collections/places.js
@@ -1,0 +1,8 @@
+
+K.extend({
+	findPlacesByCategory: function(cat) {
+		return Places.find({
+			cats: cat
+		}, K.filters.placeItem);
+	}
+});